feat: update portfolio details and enhance project components
Change the website title to "KSHITIJ NARAYAN KULKARNI" and update the favicon. Add the `react-icons` package for improved icon usage. Refactor project details in the Projects component, including renaming projects and updating their descriptions and technologies. Remove the CustomCursor component to streamline the application. Adjust layout and styling in the Hero and Skills components for better presentation.